requirements.txt
streamlit
streamlit-feedback
streamlit-extras
streamlit-chat
streamlit-option-menu
streamlit-pandas-profiling
altair
openpyxl
google-cloud-firestore
pydantic-settings
